Fix compile errors and complete build configuration
Build-breaking fixes:
- Types.h: std::fabsf -> std::fabs (non-standard MSVC identifier)
- Decrypt.cpp: Remove Decrypt_UWorld_Internal (invalid pseudo-goto syntax,
  goto across variable init C2362). Only Decrypt_UWorld_Safe is used.
- Font.h: Add #pragma once and make font array static to prevent
  LNK2005 duplicate symbol on multi-TU inclusion

Build configuration:
- Add Debug|x64 include/library paths (was only Release|x64)
- Add Debug|x64 C++20, PCH, and ForcedIncludeFiles settings
- Add link libraries (d3d11, dxgi, dwmapi, ws2_32) to both configs
- Simplify PCH NotUsing/Create to apply to all configs (not just Release)
- Switch include paths from hardcoded C:\Users\asdf to $(ProjectDir)
- Add .vcxproj.filters for Visual Studio solution explorer organization

+++ Valorant-DMA/Source/Misc/Types.h
@@ -90,7 +90,7 @@ struct alignas(16) FTransform {
         float rx = XMVectorGetX(rot), ry = XMVectorGetY(rot);
         float rz = XMVectorGetZ(rot), rw = XMVectorGetW(rot);
         float lenSq = rx * rx + ry * ry + rz * rz + rw * rw;
-        if (std::fabsf(lenSq - 1.0f) > 0.001f) return false;
+        if (std::fabs(lenSq - 1.0f) > 0.001f) return false;
 
         float sx = XMVectorGetX(scale), sy = XMVectorGetY(scale), sz = XMVectorGetZ(scale);
         if (sx <= 0.0f || sy <= 0.0f || sz <= 0.0f) return false;
